# Make writing records faster