Instead of constructing the fiona records in memory, lazily create them as you're writing (using a generator).
This gave me a 35% speedup for `collater.collate`. (For the LA dataset, from 379s (98s transforming + 281s writing) to 249s)